+++ src/types/Image.ts
@@ -1,5 +1,7 @@
 export interface RawImageItem {
   src: string
+  assetId: string
+  fileName: string
   width?: number
   heigth?: number
 }